Organize components into folders in lib
