+++ src/lib/components/Canvas/Canvas.svelte
