+++ src/lib/components/ResponsiveSVG/ResponsiveSVG.svelte
